Fixing test: of course, libmysql won't care about open_basedir, but mysqlnd should

+++ ext/mysql/tests/mysql_query_load_data_openbasedir.phpt
@@ -4,6 +4,10 @@ LOAD DATA INFILE - open_basedir
 <?php
 require_once('skipif.inc');
 require_once('skipifconnectfailure.inc');
+require_once("connect.inc");
+
+if (!$IS_MYSQLND)
+	die("skip mysqlnd only, libmysql does not know about open_basedir restrictions");
 
 if (file_exists('./simple.csv') && !unlink('./simple.csv'))
 	die("skip Cannot remove previous CSV file");
@@ -14,7 +18,6 @@ if (!$fp = fopen('./simple.csv', 'w'))
 fclose($fp);
 @unlink('./simple.csv');
 
-require_once("connect.inc");
 if ($socket == "" && $host != NULL && $host != 'localhost' && $host != '.') {
 	/* could be a remote TCP/IP connection. LOCAL INFILE may not work */
 	if (gethostbyaddr($host) != gethostname()) {
